Add files via upload
Configure your project to use the above eslint settings and fix all the warnings.
Add an endpoint to your app for the exercise calculator. It should be used by doing a HTTP POST request to endpoint exercises with the input in the request body. If the body of the request is not of the right form, response with proper status code and error message is given.