Add files via upload
Initialise a new backend project that will work with the frontend. Configure eslint and tsconfig with the same configurations as proposed in the material. Define an endpoint that answers to HTTP GET requests to route /ping.

The project should be runnable with npm scripts, both in development mode and, as compiled code, in production mode.