Add files via upload
Create a new Create React App with TypeScript, and set up eslint for the project similarly to how we just did.
All data is still kept in the App component, which passes all necessary data to each component as props. Be sure to add type declarations for each component's props!

The Header component should take care of rendering the name of the course. Content should render the names of the different parts and the amount of exercises in each part, and Total should render the total sum of exercises in all parts.